requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-frontmatter==1.0.0
markdown==3.5.1
python-dateutil==2.8.2
Jinja2==3.1.2
PyYAML==6.0.1